requirements.txt
python-telegram-bot==20.*
pandas
openpyxl
python-dotenv
